Added some transitions

+++ app/src/main/java/model_weather/Main.java
@@ -11,7 +11,7 @@ public class Main {
     private Double temp;
     @SerializedName("pressure")
     @Expose
-    private Integer pressure;
+    private Double pressure;
     @SerializedName("humidity")
     @Expose
     private Integer humidity;
@@ -37,7 +37,7 @@ public class Main {
      * @param temp
      * @param tempMin
      */
-    public Main(Double temp, Integer pressure, Integer humidity, Double tempMin, Double tempMax) {
+    public Main(Double temp, Double pressure, Integer humidity, Double tempMin, Double tempMax) {
         super();
         this.temp = temp;
         this.pressure = pressure;
@@ -69,7 +69,7 @@ public class Main {
      * @return
      *     The pressure
      */
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
@@ -78,7 +78,7 @@ public class Main {
      * @param pressure
      *     The pressure
      */
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
